Refactored profiling code
