+++ CMakeLists.txt
@@ -167,7 +167,6 @@ else ()
     add_subdirectory(src/lambda)
     add_subdirectory(src/lambda_func)
     add_subdirectory(src/memory)
-    add_subdirectory(src/prof)
     add_subdirectory(src/proto)
     add_subdirectory(src/redis)
     add_subdirectory(src/runner)
