+++ func/CMakeLists.txt
@@ -93,6 +93,7 @@ add_subdirectory(errors)
 add_subdirectory(demo)
 add_subdirectory(sgd)
 add_subdirectory(dynlink)
+add_subdirectory(polybench)
 
 # Add python build in everything but the lambda funcs
 if (FAASM_BUILD_TYPE STREQUAL "lambda-func")
